yeah yeah

+++ Common/src/main/java/uy/edu/cei/Obligatorio/Common/Controller/UsuarioController.java
@@ -7,7 +7,7 @@ import java.util.List;
 import uy.edu.cei.Obligatorio.Domain.UsuarioModel;
 
 public interface UsuarioController extends Remote {
-	public void verificarUsuario(String usuario, String contra,int id) throws RemoteException;
+	public void verificarUsuario(String usuario, String contra,long id) throws RemoteException;
 	public UsuarioModel agregarUsuario(String usuario,String contra) throws RemoteException;
 	public List<UsuarioModel> ListaUsuarios() throws RemoteException;
 }
